+++ src/MediaLibrary/Service/MediaIdParserServiceInterface.php
@@ -1,0 +1,13 @@
+<?php
+
+/**
+ * Copyright © OXID eSales AG. All rights reserved.
+ * See LICENSE file for license details.
+ */
+
+namespace OxidEsales\WysiwygModule\MediaLibrary\Service;
+
+interface MediaIdParserServiceInterface
+{
+    public function parseMediaIdsFromContent(string $content): array;
+}